requirements.txt
python-dotenv
win10toast
pyttsx3
pytz
genshin @ git+https://github.com/thesadru/genshin.py